requirements.txt
maturin==1.3.1,
scallopy==0.2.1,
pykeen==1.10.1,
torch==2.1.0,
torch-geometric==2.4.0,
torch-scatter==2.1.0,
torch-sparse==0.6.18,
ltntorch==1.0.1,
jupyter==1.0.0,
matplotlib==3.8.1
wandb==0.16.0
transformers==4.32.1
pyyaml==6.0.1